% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ff_table.R
\name{add_ff_table}
\alias{add_ff_table}
\title{Failure count}
\usage{
add_ff_table(ggaux, multi_surv_reg)
}
\arguments{
\item{ggaux:}{a plot reulting from \code{plot_predict_multi_surv_reg}.}

\item{multi_surv_reg:}{List containing at least the element 'ff_table'.}
}
\value{
A ggplot.
}
\description{
add failure counts to plot
}
\details{
Add geom_text containing the failure, suspension and total counts.
}
\examples{

}
