% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_target_point.R
\name{add_target_point}
\alias{add_target_point}
\title{Target point}
\usage{
add_target_point(ggaux, models_list, time_target = 0, surv_target = 1,
  legend_text = paste0("Target: ", surv_target * 100, "@", time_target),
  p_size = 5, p_shape = 18, t_size = 3, adj_axis = TRUE,
  failure_rate_plot = FALSE)
}
\arguments{
\item{time_target, }{surv_target: The target 'time' and 'survival'}

\item{p_size, }{p_shape, t_size: point size, point shape and text size.}

\item{adj_axis:}{if \code{TRUE}, plot will be adjusted to show all estimated survival for the models.}

\item{failure_rate_plot:}{if \code{TRUE}, the estimates of failure rate (1-survival) will be ploted instead. This is usefull for weibull paper, mostly.}

\item{ggaux:}{a \code{ggplot} object.}

\item{legend_text:}{If not specified, the legend will be in the format "surv_target@time_target"}

\item{models_list:}{An object or a \code{list} of objects returned from code{multi_surv_reg()}}
}
\value{
A \code{ggplot}
}
\description{
Add target point and estimated survival of the models for the givem target.
}
\examples{

}
