% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_km_estimates_from_flexsurvreg.R
\name{extract_km_estimates_from_flexsurvreg}
\alias{extract_km_estimates_from_flexsurvreg}
\title{Kaplan-Meyer survival estimates}
\usage{
extract_km_estimates_from_flexsurvreg(m_flexsurvreg, ic = FALSE)
}
\arguments{
\item{m_flexsurvreg:}{an object or list of objects of class \code{flexsurvreg}. If m_flexsurvreg is a list the function will extract the data of only the first element.}

\item{ic:}{If \cpde{TRUE} returns the confidence interval}
}
\value{
: A \code{data.framme} containing the Kaplan-Meyer survival estimates.
}
\description{
Extracts the Kaplan-Meyer survival estimates from a 'flexsurvreg' object
}
\examples{

}
