% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_Rfilter_to_title_str.R
\name{from_Rfilter_to_title_str}
\alias{from_Rfilter_to_title_str}
\title{Title of the function}
\usage{
from_Rfilter_to_title_str(rule)
}
\arguments{
\item{rule:}{A \code{character} filter rule.}
}
\value{
a \code{character}
}
\description{
converts an R filter \code{character} to a more human readable text
}
\examples{
from_Rfilter_to_title_str('country \%in\% c("BRA", "GER", "JPN") & age >= 30')

}
