% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gof_measures.R
\name{get_gof_measures}
\alias{get_gof_measures}
\title{Parametrical and non parametrical fit compare.}
\usage{
get_gof_measures(KM_fit, m_flexsurvreg)
}
\arguments{
\item{KM_fit:}{a \code{survfit} object}

\item{m_flexsurvreg:}{a \code{flexsurvreg} object}
}
\value{
: A \code{data.table} with the comparison results.
}
\description{
Compare flexsurv and survfit
}
\details{
# Returns a \code{data.frame} with the columns
fit_mse: The mean squared error of the linear model %$S_{km} ~ S_{model$%
mse: average of the squared difference between %$S_{km} - S_{model$%
rho_KM_flex: correlation between %$S_{km} $% and %$S_{model$%
loglik: log likelihood of the model
p_ks_test: p-value of the Kolmogorov-Smirnov test \code{ks.test()}
}
\examples{

}
