% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters_flexsurvreg.R
\name{get_parameters_flexsurvreg}
\alias{get_parameters_flexsurvreg}
\title{Get distribution parameters}
\usage{
get_parameters_flexsurvreg(model_flexsurvreg, name = FALSE)
}
\arguments{
\item{model_flexsurvreg:}{an object of class 'flexsurvreg'.}

\item{name:}{\code{logical} indicating whether to return name of the distribution or not.}
}
\description{
get parameters from model 'flexsurvreg'
}
\details{
get parameters from model 'flexsurvreg'
}
\examples{

}
