% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_out_table.R
\name{in_out_table}
\alias{in_out_table}
\title{IN & OUT Table}
\usage{
in_out_table(data, reference_month, RM_size, status_col = "status",
  assembly_col, repair_col, no_fail_dist = 2)
}
\arguments{
\item{status_col, }{assembly_col, repair_col: \code{character} status, assembly and repair columns.}

\item{data:}{A \code{data.frame}}

\item{no_fail_dist:}{number of months to concatenate the column containing the distribution of failures along the assembly month. Usefull to see if the recent failures are of individuals recently born.}

\item{reference_month:}{reference month}

\item{RM_size:}{RM size.}
}
\description{
The table for the IN & OUT analysis.
}
\examples{

}
