% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_to_SQLfilter.R
\name{interval_to_SQLfilter}
\alias{interval_to_SQLfilter}
\title{Title of the function}
\usage{
interval_to_SQLfilter(field_name = "", values = c(""))
}
\arguments{
\item{field_name:}{the name of the field in the data base}

\item{values:}{the values to be filtered}
}
\value{
A \code{character} containing the statement.
}
\description{
makes an SQL interval filter statement
}
\details{
A filter statement for interval limit.
}
\examples{
interval_to_SQLfilter(field_name = 'age', c('>=18','<20'))

}
