% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_Rfilter.R
\name{list_to_Rfilter}
\alias{list_to_Rfilter}
\title{Rule string}
\usage{
list_to_Rfilter(rules_list = list(rule1 = 1:3, rule2 = ">2"))
}
\arguments{
\item{rules_list:}{Named \code{list} containing the values of the filter.}
}
\value{
A \code{character} containing the rule.
}
\description{
make rule string for dplyr::filter() taking from list
}
\examples{
rules = list(
  country = c('BRA', 'GER', 'JPN'),
  grade = c('A', 'B', 'C'),
  age = c('>10', '<25')
  
)
list_to_Rfilter(rules)

}
