% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_SQLfilter.R
\name{list_to_SQLfilter}
\alias{list_to_SQLfilter}
\title{Title of the function}
\usage{
list_to_SQLfilter(rules_list = list(rule1 = 1:3, rule2 = ">2"))
}
\arguments{
\item{rules_list:}{Named \code{list} containing the values of the filter.}
}
\value{
A \code{character} containing the rule.
}
\description{
Description of the function
}
\examples{
rules = list(
  country = c('BRA', 'GER', 'JPN'),
  grade = c('A', 'B', 'C'),
  age = c('>10', '<25')
  
)
list_to_SQLfilter(rules)

}
