% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_status_from_rule.R
\name{make_status_from_rule}
\alias{make_status_from_rule}
\title{Status Rule}
\usage{
make_status_from_rule(dfr, rule)
}
\arguments{
\item{dfr:}{a \code{data.frame}}

\item{rule:}{A \code{character} (or a syntactically valid \code{R} expression_ containing the expresion that will determine if an observation is a failure.}
}
\description{
Make status column from a given rule
}
\examples{

}
