% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_surv_reg.R
\name{multi_surv_reg}
\alias{multi_surv_reg}
\title{Best suvreg adjust}
\usage{
multi_surv_reg(surv_data, time_col = "time", status_col = "status",
  surv_formula = "~ 1", dist_list = "", ...)
}
\arguments{
\item{dist_list:}{A vector containing a list of distributions to fit survival models and pick the best.}

\item{status_col:}{The ame of the column containing the status}

\item{surv_data:}{a \code{data.frame}}

\item{surv_formula:}{A \code{formula} for the regression.}

\item{time_col:}{name of the column containing the time}

\item{...:}{further arguments for flexsurvreg()}
}
\value{
A \code{list} containing the best survreg fits
}
\description{
Finds best suvreg adjust
}
\details{
All distributions in the \code{dist_list} are used as link functions for the survival regresion model. The models are returned in the \code{models} slot of the result, ordered by best fit. The slot \code{erros} contains the fits that had erros. \code{Adjust_table} contains the goodness of fit of the models. \code{ff_table} contains the faulire frequency table. \code{KM_fit} has the Kaplan-Meier fit.
}
\examples{

da = system.file("extdata", "example.csv", package = "reliabilitytools")
da = read.csv2(da)

da_status = da \%>\%#'   
  mutate(status = as.integer(failure_cause_group == 560)) # cause of failure
  
  
# -- prepare life data
da_prepared = prepare_life_times(da_status
                                 , indiv_col = 'id'
                                 , status_col = 'status'
                                 , obs_time_col = 'cycles')
# -- suvival models
surv_model = multi_surv_reg(da_prepared
                            , time_col = 'elapsed'
                            , status_col = 'status')

# -- plot
plot_life_curve(surv_model$models,
                title_aux = '\\nFailure cause: 560')

}
