% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_surv_spline.R
\name{multi_surv_spline}
\alias{multi_surv_spline}
\title{Survival Spline fit}
\usage{
multi_surv_spline(surv_data, time_col = "time", status_col = "status",
  surv_formula = "~ 1", scale_list = "", knots = 3, ...)
}
\arguments{
\item{knots:}{Number of knots.}

\item{scale_list:}{A vector containing a list of scale arguments for \code{flexsurvspline}.}

\item{status_col:}{Name of the column containing the status.}

\item{surv_data:}{a \code{data.frame}}

\item{surv_formula:}{a \code{formula}}

\item{time_col:}{Name of the column containing the time.}

\item{...:}{further arguments for \code{flexsurvspline}.}
}
\value{
The models are returned in the \code{models} slot of the result, ordered by best fit. The slot \code{erros} contains the fits that had erros. \code{Adjust_table} contains the goodness of fit of the models. \code{ff_table} contains the faulire frequency table. \code{KM_fit} has the Kaplan-Meier fit.
}
\description{
Finds best spline suvreg fit
}
\examples{

}
