% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_from_excel.R
\name{params_from_excel}
\alias{params_from_excel}
\title{Paramenters from excel}
\usage{
params_from_excel(file_path = "", sheet = "params",
  end_period = as.Date(Sys.time()), classification_columns = NA,
  filter_columns = NA, ...)
}
\arguments{
\item{classification_columns:}{: \code{character} vector containing the names of the columns tha will be used in the classification (failure/suspenstion)}

\item{end_period:}{upper year month limit for the RM.}

\item{file_path:}{Path to the excell file containing the parameters}

\item{filter_columns:}{\code{character} vector containing the names of the columns that will be used to filter the population.}

\item{sheet:}{sheet to read from.}

\item{...:}{further arguments for \code(read_excel()).}
}
\description{
create paramenters from excel table
}
\details{
makes a list containing the parameters for the RM
This function was coded to optimize the RM reports generation.
The algorithm for RM is described below:
    1) Separate the population in risk
    2) Classify the statuses as 'event' or 'suspension'
    3) Prepare times (calculate elapsed times)
    4) Estimate life curves.

Often, one wants to estimate several curves for different failures causes for the same 
Population Under Risk (PUR). So there is no need to separate the population every time,
e.g  For the population market_type = FH, the Step 1) needs be executed only once.
Once the PUR data is acquired, one can estimate life curves for 'Turbo' , 
'Gear Box' and 'Valves'.
}
\examples{

}
