% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_failure_prob_curve_overlap.R
\name{plot_failure_prob_curve_overlap}
\alias{plot_failure_prob_curve_overlap}
\title{Overlaped failure rate curves}
\usage{
plot_failure_prob_curve_overlap(models_groups,
  group_names = names(models_groups), title_aux = "",
  x_label = "Mileage(km)", y_label = "Failure Prob.", add_curves = TRUE,
  best_model = 1)
}
\arguments{
\item{x_label, }{y_label:  X and Y axis labels}

\item{add_curves:}{if true, the distribution curves will be added}

\item{best_model:}{The rank of model to be used in the plot or the name fo the distribution. If best_model = 2, the second best model will be used in the plot}

\item{group_names:}{Names for the curves}

\item{models_groups:}{A \code{list} of objects containing the element 'models' (which is a list of flexsurvreg')}

\item{title_aux:}{complement to the title}
}
\value{
A ggplot of the plots overlaped.
}
\description{
Plot two or more failure rate curves on top of each other
}
\examples{

}
