% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_life_curve.R
\name{plot_life_curve}
\alias{plot_life_curve}
\title{Life curve}
\usage{
plot_life_curve(m_flexsurvreg, title_aux = "", x_label = "time",
  y_label = "reliability", ylim = c(0, 1), xlim = NA, line_width = 1.5)
}
\arguments{
\item{x_label, }{y_label: labels for Y and X axis}

\item{m_flexsurvreg:}{An object of class flexsurvreg, or a list of objects 'flexsurvreg'. If a list of models is given, the plot will have all the curves}

\item{title_aux:}{a complement to the title}

\item{xlim:}{limits for x-axis. A blank point is added to to the plot, otherwise the life curves are not fully drawn.if xlim is vector of length 1, this will be the upper limmit.}

\item{ylim:}{limits for y-axis. if ylim is vector of length 1, this will be the lower limmit}
}
\value{
A ggplot of the life curve.
}
\description{
Plot life curve
}
\examples{


}
