% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_life_curve_multi_surv_reg.R
\name{plot_life_curve_multi_surv_reg}
\alias{plot_life_curve_multi_surv_reg}
\title{Best Life curve}
\usage{
plot_life_curve_multi_surv_reg(multi_surv_reg, pred_times = c(50, 100, 120,
  150, 200) * 1000, conf_int = FALSE, aux_title = "", ylim = c(0, 1),
  time_scale = 1000, best_model = 1)
}
\arguments{
\item{aux_title:}{Complement to title}

\item{best_model:}{The rank of model to be used in the plot. If best_model = 2, the second best model will be used in the plot}

\item{conf_int:}{Add confidence interval if \code{TRUE}}

\item{multi_surv_reg:}{List containing at least the alement "models" that has a list of \code{flexsurvreg} models.}

\item{pred_times:}{Prediction times.}

\item{time_scale:}{a number to devide the time so it rescales and the legend looks better}

\item{ylim:}{Y axis limits}
}
\value{
A ggplot RM life curve of the best model.
}
\description{
Plot life curve from best model
}
\examples{

}
