% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_life_curve_overlap.R
\name{plot_life_curve_overlap}
\alias{plot_life_curve_overlap}
\title{Overlaped life curves}
\usage{
plot_life_curve_overlap(models_groups, group_names = names(models_groups),
  title_aux = "", x_label = "Mileage (km)", y_label = "Reliability (\%) ",
  ylim = c(0, 1), add_curves = TRUE, best_model = 1,
  add_interpolation_line = FALSE)
}
\arguments{
\item{add_curves:}{if \code{TRUE}, the distribution curves will be added.}

\item{add_interpolation_line:}{if \code{TRUE}, adds a line connecting the points.}

\item{best_model:}{The rank of model to be used in the plot or the name fo the distribution. If best_model = 2, the second best model will be used in the plot}

\item{group_names:}{Names for the curves}

\item{models_groups:}{A \code{list} of objects containing the element 'models' (which is a list of flexsurvreg')}

\item{title_aux:}{complement to the title}

\item{ylim:}{Y axis limit.}
}
\value{
A ggplot of the plots overlaped.
}
\description{
Plot two or more life curves on top of each other
}
\examples{

}
