% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predict_multi_surv_reg.R
\name{plot_predict_multi_surv_reg}
\alias{plot_predict_multi_surv_reg}
\title{RM prediction curves}
\usage{
plot_predict_multi_surv_reg(multi_surv_reg, time_scale = 1,
  col_legend = "time", aux_title = "", ylim = c(0, 100.05),
  x_grid_step = 50000, ic_area = TRUE)
}
\arguments{
\item{aux_title:}{A title complement.}

\item{col_legend:}{a title for the legend.}

\item{ic_area:}{\code{logical} add confidence interval \code{ribbon} of the max time (lowest line).}

\item{multi_surv_reg:}{A \code{multi_surv_reg object}}

\item{time_scale:}{a number to devide the time so it rescales and the legend looks better}

\item{x_grid_step:}{The spacing between each x grid line.}

\item{ylim:}{y axix limits}
}
\value{
A ggplot of the RM  prediction data.frame
}
\description{
Plot RM prediction curves from a data.frame ####
}
\examples{

}
