% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_best_model.R
\name{predict_best_model}
\alias{predict_best_model}
\title{Best Fit Prediction}
\usage{
predict_best_model(m_flexsurvreg, pred_times = c(50, 100, 120, 150, 200) *
  1000, conf_int = FALSE, best_model = 1, ...)
}
\arguments{
\item{best_model:}{The rank of model to be used in the plot. If best_model = 2, the second best model will be used in the plot. Can be the name of the distribution of the model}

\item{conf_int:}{If \code{TRUE} then the confidence intervals will be calculated.}

\item{m_flexsurvreg:}{list containing at least the alement "models" that has a list of \code{flexsurvreg} models. Or an object of class \code{flexsurvreg}}

\item{pred_times:}{the list of times to predict to.}

\item{...:}{further argumentrs for \code{summary()}}
}
\description{
Predicts the survival/reliability of the given times for the model
}
\details{
Returns a data.frame with the predicted survival/reliability for  range of life times (resulting from a \code{multi_surv_reg()}). If the object is a model itself then the function does not search for the best model.
}
\examples{

}
