% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_best_model_multi_surv_reg.R
\name{predict_best_model_multi_surv_reg}
\alias{predict_best_model_multi_surv_reg}
\title{Bulk best Model predict}
\usage{
predict_best_model_multi_surv_reg(multi_surv_reg, pred_times = c(50, 100, 120,
  150, 200) * 1000, conf_int = FALSE, best_model = 1)
}
\arguments{
\item{best_model:}{The rank of model to be used in the plot. If best_model = 2, the second best model will be used in the plot. Can be the name of the distribution of the model}

\item{conf_int:}{if true then the confidence intervals will be calculated.}

\item{multi_surv_reg:}{a named list containing at least the alement "models" that has a list of "flexsurvreg" models}
}
\value{
A \code{data.frame} with the predicted survival/reliability for  range of life times (resulting from a \code{multi_surv_reg()})
}
\description{
Predicts the survival/reliability of the given times for the model on a multi_surv_reg \code{list}
}
\examples{

}
