% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_table.R
\name{prediction_table}
\alias{prediction_table}
\title{Prediction table}
\usage{
prediction_table(rel_model, avg_cost = NA, ...)
}
\arguments{
\item{avg_cost:}{a \code{numeric} containing the avarage cot per failure}

\item{rel_model:}{a \code{flexsurvreg} object}

\item{...:}{further arguments for predict_best_model()}
}
\value{
A prediction \code{data.table} of a given model
}
\description{
Prediction table
}
