% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepares_indiv_life_times.R
\name{prepare_indiv_life_times}
\alias{prepare_indiv_life_times}
\title{Life time prepare}
\usage{
prepare_indiv_life_times(data, indiv_col = "indiv",
  obs_time_col = "obs_time", status_col = "status")
}
\arguments{
\item{data:}{A \code{data.frame}}

\item{indiv_col:}{column containing the identification of the individual.}

\item{obs_time_col:}{Observed time of the event}

\item{status_col:}{Event (O suspension, 1 "death")}
}
\value{
A \code{data.frame} appropriately prepared for reliability analysis.
}
\description{
Preprares life times of one individual
}
\details{
The codes gets the registered times and status and transforms them into life data.
Especifications:
  -If the longest time is a failure there is no need to take suspenstions into account.
  -Only the longest observed suspension has to be considered.
  -Lifetimes equal to Zero will no be considered:
    +If an individual has failed from two causes at the same time there will be tied events.
      This will lead to events with liefetime equal to Zero, which can not be considered for the survival computations
    +Failures occouring at time Zero can not be considered for the survival calculations nor the failure cause table.
}
\examples{

}
