% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepares_life_times.R
\name{prepare_life_times}
\alias{prepare_life_times}
\title{Bulk life time preparatio.}
\usage{
prepare_life_times(data, indiv_col = "indiv", status_col = "status",
  obs_time_col = "obs_time")
}
\arguments{
\item{data:}{A \code{data.frame} to be prepared}

\item{indiv_col:}{column containing the identification of the individual.}

\item{obs_time_col:}{Observed time of the event}

\item{status_col:}{Event (O suspension, 1 "death")}
}
\value{
A \code{data.frame} conatining the prepared data.
}
\description{
Prepare life times for each individual in the dataset
}
\details{
The codes calls prepare_indiv_life_times for each individual.
}
\examples{

da = system.file("extdata", "example.csv", package = "reliabilitytools")
da = read.csv2(da)

da \%>\%
  mutate(status = as.integer(failure_cause_group == 560)) \%>\%
  prepare_life_times(indiv_col = 'id'
                     , status_col = 'status'
                     , obs_time_col = 'cycles')


}
