% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_summary.R
\name{reliability_summary}
\alias{reliability_summary}
\title{Reliability summary information}
\usage{
reliability_summary(rel_analysis, reference_month, pareto_col = NA,
  rel_quantiles = c(0.05, 0.1))
}
\arguments{
\item{reference_month:}{Reference month.}

\item{rel_analysis:}{a \code{list} containing the elements \code{data} and \code{models}, which are \code{lists}.}

\item{rel_quantiles:}{\code{numeric} probabilities to estimate time quantiles for.}
}
\value{
a summary  \code{list} containing the most relevant information of the analysis.
}
\description{
A summary of reliability information
}
\examples{

}
