% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_data_separate_SQlite.R
\name{rm_data_separate_SQlite}
\alias{rm_data_separate_SQlite}
\title{Title of the function}
\usage{
rm_data_separate_SQlite(data_base = "", RM_size = 6, current_month,
  population_filter = NA, select_statement = NA,
  assemb_col = "DWASSD_yearmon", repair_col = "FPAYDT_yearmon")
}
\arguments{
\item{assemb_col, }{repair_col: assembly and repair columns .}

\item{current_month:}{a \code{character} object of Month of the given RM.}

\item{data_base:}{A string containing the SQLite3 data_base}

\item{population_filter:}{A filter "WHERE" statement.}

\item{RM_size:}{the size, in momths, of the RM. Typically 6, 9, 12, 15, 18, 21 or 24}

\item{select_statement:}{A \code{character} containing a "SELECT" statement.}
}
\value{
A \code{tibble}
}
\description{
Description of the function
}
\details{
Makes the RMx data set of the given month.
}
\examples{

}
