% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_flexsurvreg.R
\name{safe_flexsurvreg}
\alias{safe_flexsurvreg}
\title{Safe version of flexsurv::flexsurvreg()}
\usage{
safe_flexsurvreg(...)
}
\description{
Safe version of flexsurv::flexsurvreg()
}
\examples{

}
