% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_flexsurvspline.R
\name{safe_flexsurvspline}
\alias{safe_flexsurvspline}
\title{Safe version of flexsurv::flexsurvspline()}
\usage{
safe_flexsurvspline(...)
}
\description{
Safe version of flexsurv::flexsurvspline()
}
\examples{

}
