% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexsurvreg_NULL_methods.R
\name{summary.flexsurvreg_NULL}
\alias{summary.flexsurvreg_NULL}
\title{Title of the function}
\usage{
\method{summary}{flexsurvreg_NULL}(x, t = 1, ci = FALSE, ...)
}
\arguments{
\item{ci:}{confidence interval \code{logical}}

\item{t:}{time}

\item{x:}{obj}
}
\description{
Prediction method for the flexsurvreg_NULL class
}
\details{
This method is necessary because when there is no data , the survival curve will not be made. And when the script needs to predict the survival for this data set, it has to return NA.
}
\examples{

}
