% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_to_sqlfilter.R
\name{vector_to_sqlfilter}
\alias{vector_to_sqlfilter}
\title{AND SQL filter 
makes an 'AND' filter statement}
\usage{
vector_to_sqlfilter(field_name = "", values = c(""))
}
\arguments{
\item{field_name}{The name of the fild in the data base.}

\item{values}{The values to be filtered}
}
\value{
A \code{character} containing the SQL statement.
}
\description{
AND SQL filter 
makes an 'AND' filter statement
}
\examples{
vector_to_sqlfilter(field_name = 'country',
values = c('BRA', 'GER', 'JPN'))

}
